XP-3478 Delete dialog - Show spinner while fetching child items - Moved mask code to a common for delete and publish parent - DependantItemsDialog - Simplified a little of code in DeletedDialog
